# openapi/openapi.yaml
basePath: /
definitions:
...
  models.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.APIError'
      success:
        type: boolean
    type: object
...
  models.BucketListResponse:
    properties:
      buckets:
        items:
          $ref: '#/definitions/models.BucketInfo'
        type: array
      count:
        type: integer
    type: object
...
  models.CreateBucketRequest:
    properties:
      name:
        type: string
      region:
        type: string
    required:
    - name
    type: object
...
  models.ObjectDeleteResponse:
    properties:
      bucket:
        type: string
      deleted:
        type: boolean
      key:
        type: string
    type: object
...
  models.ObjectUploadResponse:
    properties:
      bucket:
        type: string
      content_type:
        type: string
      etag:
        type: string
      key:
        type: string
      size:
        type: integer
    type: object
...
  models.PresignedURLResponse:
    properties:
      bucket:
        type: string
      expires_in:
        description: in seconds
        type: integer
      key:
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    REST API for managing Garage distributed object storage system
    This API provides endpoints for managing buckets, objects, users, and cluster operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Garage UI API
  version: 0.1.0
paths:
  /api/v1/buckets:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all buckets in the Garage storage system with
        object count and size
      operationId: listBuckets
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of buckets
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.BucketListResponse'
              type: object
        "500":
          description: Failed to list buckets
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: List all buckets
      tags:
      - Buckets
    post:
      consumes:
      - application/json
      description: Creates a new bucket in the Garage storage system
      parameters:
      - description: Bucket creation payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.CreateBucketRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Bucket created successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    bucket:
                      type: string
                    message:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid request body or bucket name is required
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "409":
          description: Bucket already exists
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "500":
          description: Failed to create bucket
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new bucket
      tags:
      - Buckets
  /api/v1/buckets/{bucket}/directories:
    post:
      consumes:
      - application/json
      description: Creates a zero-byte object whose key ends with "/" so that S3 clients
        display it as an empty folder.
      parameters:
      - description: Name of the bucket
        in: path
        name: bucket
        required: true
        type: string
      - description: Directory key (must end with '/')
        in: body
        name: request
        required: true
        schema:
          properties:
            key:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Directory created
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ObjectUploadResponse'
              type: object
        "400":
          description: Invalid request parameters
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "500":
          description: Failed to create directory
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Create directory in bucket
      tags:
      - Objects
  /api/v1/buckets/{bucket}/object:
    delete:
      consumes:
      - application/json
      description: Deletes an object stored in the specified bucket
      operationId: deleteObject
      parameters:
      - description: Name of the bucket containing the object
        in: path
        name: bucket
        required: true
        type: string
      - description: Exact object key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted the object
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ObjectDeleteResponse'
              type: object
        "400":
          description: Bucket name and object key are required
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "404":
          description: Object not found
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "500":
          description: Failed to delete object
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Delete object from bucket
      tags:
      - Objects
    get:
      consumes:
      - application/json
      description: Retrieves an object stored in the specified bucket
      operationId: getObject
      parameters:
      - description: Name of the bucket containing the object
        in: path
        name: bucket
        required: true
        type: string
      - description: Exact object key
        in: query
        name: key
        required: true
        type: string
      - description: Set to true to download the object as an attachment
        in: query
        name: download
        type: boolean
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Successfully retrieved the object
          schema:
            type: file
        "206":
          description: Partial object content
          schema:
            type: file
        "400":
          description: Bucket name and object key are required
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "404":
          description: Object not found
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "416":
          description: Requested range is not satisfiable
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Get object from bucket
      tags:
      - Objects
    head:
      description: Returns the same object representation headers as GET without transferring
        the object body.
      operationId: headObject
      parameters:
      - description: Name of the bucket containing the object
        in: path
        name: bucket
        required: true
        type: string
      - description: Exact object key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Head an object
      tags:
      - Objects
  /api/v1/buckets/{bucket}/object/metadata:
    get:
      consumes:
      - application/json
      description: Retrieves metadata information about an object without downloading
        the actual content
      operationId: getObjectMetadata
      parameters:
      - description: Name of the bucket containing the object
        in: path
        name: bucket
        required: true
        type: string
      - description: Exact object key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved object metadata
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ObjectInfo'
              type: object
        "400":
          description: Bucket name and object key are required
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "404":
          description: Object not found
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Get object metadata
      tags:
      - Objects
  /api/v1/buckets/{bucket}/object/presign:
    get:
      consumes:
      - application/json
      description: Generates a pre-signed URL that allows temporary access to the
        specified object
      operationId: getObjectPresignedURL
      parameters:
      - description: Name of the bucket containing the object
        in: path
        name: bucket
        required: true
        type: string
      - description: Exact object key
        in: query
        name: key
        required: true
        type: string
      - description: 'Expiration time in seconds for the pre-signed URL (default:
          3600 seconds)'
        in: query
        name: expires_in
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully generated pre-signed URL
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PresignedURLResponse'
              type: object
        "400":
          description: Invalid request parameters
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "404":
          description: Object not found
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "500":
          description: Failed to generate pre-signed URL
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Get pre-signed URL for object
      tags:
      - Objects
  /api/v1/buckets/{bucket}/object/preview-url:
    get:
      description: Returns a relative URL carrying a short-lived token that authorizes
        streaming this object. Media elements cannot send an Authorization header,
        so the token rides in the URL instead.
      operationId: getObjectPreviewURL
      parameters:
      - description: Name of the bucket containing the object
        in: path
        name: bucket
        required: true
        type: string
      - description: Exact object key
        in: query
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Preview URL minted
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.PreviewURLResponse'
              type: object
        "400":
          description: Bucket name and object key are required
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
        "500":
          description: Failed to mint the preview token
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: Get a tokenized preview URL for an object
      tags:
      - Objects
...
  /auth/config:
    get:
      description: Returns the current auth configuration (admin and/or OIDC)
      operationId: getAuthConfig
      produces:
      - application/json
      responses:
        "200":
          description: Auth config
          schema:
            $ref: '#/definitions/models.AuthConfigResponse'
        "500":
          description: Configuration unavailable
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Get authentication configuration
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...